Harden the API: [[nodiscard]] on every entry point, clang-only warning policy
This library is vendored into ClickHouse, and integrating it surfaced a class of
bug the API made easy to write and impossible to detect. Every encode/decode
returns the number of bytes written or consumed, and that count is the *only* way
a caller can find where the next block begins. Dropping it silently
desynchronises the caller's cursor and produces plausible-looking garbage far
away from the mistake. Two such bugs were found during the ClickHouse migration,
one of them pointer self-assignment UB.

Compressed output is bit-for-bit unchanged, and so is the generated code: over 72
(data shape x entry style x width x block size) combinations, callgrind reports
the same retired-instruction count as before this commit, exactly, in all 72.

## [[nodiscard]] on all 72 entry points

Ignoring the return value is now a compile error unless the caller opts out with
`std::ignore = ...`, which documents at the call site that the length is known by
other means. This is the whole of the fix; the diff to include/abpfor.h is 72
declarations with `[[nodiscard]]` prepended and nothing else changed.

Return types are untouched -- every count stays `size_t`. An earlier version of
this change narrowed the single-block functions to `uint32_t`, on the theory that
a bounded output deserves a bounded type. That was wrong on all three counts:

  * It buys no safety. `[[nodiscard]]` is what catches a dropped count, and it
    works the same at either width.
  * It buys no clarity. The bound that actually matters belongs to the consumer
    (ClickHouse stores block lengths in a `UInt32` wire format), not to this
    library, and the consumer already asserts it. Moving that constraint into the
    producer's signature points the dependency the wrong way.
  * It costs. On x86-64 the upper half of a 32-bit return register is not
    guaranteed, so `p += f()` with a `uint32_t`-returning `f` and a pointer `p`
    emits a `movl %eax, %eax`. Callers pay one per block; narrowing the internal
    layers too made this library pay one per block inside its own stream loops,
    for counts that never leave the library. Measured: +0.108% instructions on
    the stream paths from the internal narrowing alone, +0.07% on block-loop
    paths from the public boundary.

It also actively hid something. ClickHouse builds with -Wshorten-64-to-32
-Werror; a `uint32_t` return let 34 call sites assign a byte count into a `UInt32`
field with no cast and no warning. With `size_t` those narrowings are explicit
again, next to the bound that justifies them.

## Warning policy mirroring ClickHouse's

CMakeLists.txt now uses `-Weverything -Wpedantic` with individual warnings
disabled only where they are noise for a bit-packing codec, each carrying an
inline reason -- the same structure as ClickHouse's cmake/warnings.cmake. An
opt-in `ABPFOR_WERROR` promotes them to errors. The tree builds at zero warnings
under `-Weverything -Werror` in both Release and Debug; all 15 tests pass.

Two source changes came out of this:

  * The two `unpack_delta_b{8,16}_emit` fold expressions sequenced their
    per-element statements with the comma operator, which -Wcomma flags as
    indistinguishable from a typo. They are now a lambda per element. The
    generated assembly is byte-identical, verified by diffing -S output, and by
    building baseline+lambda and HEAD-minus-lambda and confirming both match
    their counterpart instruction-for-instruction.
  * Seven implicit narrowings became explicit casts (three `uint32_t` ->
    `unsigned char` in pack_b3_loop, four around `__builtin_popcount` and
    `_mm_slli_epi32` in avx2Scatter). A cast silences the compiler whether or not
    it is correct, so tests/test_narrowing.cpp asserts each one's bound: the
    24-bit packing range, the 0..4 nibble popcount, and the 0..64 bit width.

## No runtime asserts

Per this library's stated policy -- it validates nothing, and every precondition
violation is UB rather than an error return -- there are no runtime asserts. The
six `assert(n < BS)` in api_block.h are gone; they were dead code anyway, since
NDEBUG was hardcoded into the shared compile flags. NDEBUG stays set
unconditionally, now with a comment saying why.

## GCC is rejected at configure time

GCC's -O3 loop vectoriser miscompiles the interleaved unpack path at bit widths
10/12/26/28/30 (32-bit) and 38/44 (64-bit), failing 4 tests;
-fno-tree-loop-vectorize makes them pass. That is pre-existing, not a regression
from this work. Rather than ship a library that builds but computes wrong
answers, `project()` now hard-errors on non-clang compilers. The consumer is
clang-only anyway, so this costs nothing.

## Documentation

README gains the `[[nodiscard]]` contract and why the width is `size_t`, the
calling convention (`p += decodeBlock(p, out)`, never
`end = decodeBlock(p, out) + p` -- encode wants the output base and decode the
input base, so a swapped base compiles fine and corrupts silently), the clang
requirement, the warning policy, and an output-buffer-size section: b128/b256
`decodeBlock` takes no element count and always writes a full block, so a caller
sizing the buffer to its actual element count overflows the heap. Every README
snippet was compiled and run, with roundtrip assertions on the results.

+++ src/simd/avx2_fused.h
@@ -97,12 +97,14 @@ ABPFOR_INLINE __m256i avx2Scatter(const uint32_t*& pex, unsigned mask8, unsigned
 
     // Load up to 8 contiguous residuals (may over-read, safe with padding)
     __m128i excLo = _mm_loadu_si128(reinterpret_cast<const __m128i*>(pex));
-    unsigned loCount = __builtin_popcount(loNib);
+    // popcount of a 4-bit nibble is 0..4, so the int -> unsigned narrowing is exact.
+    unsigned loCount = static_cast<unsigned>(__builtin_popcount(loNib));
     __m128i excHi = _mm_loadu_si128(reinterpret_cast<const __m128i*>(pex + loCount));
 
-    // Shift residuals left by B bits
-    __m128i excLoShifted = _mm_slli_epi32(excLo, B);
-    __m128i excHiShifted = _mm_slli_epi32(excHi, B);
+    // Shift residuals left by B bits. B is a bit width (0..64), so it always fits
+    // the int shift-count the intrinsic takes.
+    __m128i excLoShifted = _mm_slli_epi32(excLo, static_cast<int>(B));
+    __m128i excHiShifted = _mm_slli_epi32(excHi, static_cast<int>(B));
 
     // Scatter via LUT
     __m128i maskLo = _mm_load_si128(reinterpret_cast<const __m128i*>(detail::kScatterLUT8[loNib]));
@@ -114,7 +116,8 @@ ABPFOR_INLINE __m256i avx2Scatter(const uint32_t*& pex, unsigned mask8, unsigned
     // Combine into __m256i
     __m256i result = _mm256_inserti128_si256(_mm256_castsi128_si256(scatLo), scatHi, 1);
 
-    pex += loCount + __builtin_popcount(hiNib);
+    // Same 0..4 popcount bound as loCount above: the int -> unsigned cast is exact.
+    pex += loCount + static_cast<unsigned>(__builtin_popcount(hiNib));
 
     return result;
 }
